requirements.txt
numpy==2.2.6
huggingface_hub==0.35.3
hf_transfer==0.1.9
transformers==4.56.2
loguru==0.7.3
einops==0.8.1
timm==1.0.20
packaging==25.0

git+https://github.com/huggingface/diffusers.git
peft

accelerate==1.10.1
kornia==0.8.1
xformers==0.0.31
vllm==0.10.0
ray==2.49.2
requests==2.32.5
urllib3==2.5.0
idna==3.10
tokenizers==0.22.0
albumentations==2.0.8
torch-c-dlpack-ext==0.1.3

lightning==2.5.4
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
omegaconf==2.3.0
wandb==0.20.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
protobuf==5.29.5
jaxtyping==0.3.2
spconv-cu126==2.3.8
scikit-learn==1.7.0
python-multipart==0.0.20
numba==0.61.2
opencv-python-headless==4.12.0.88
rootutils==1.0.7
wrapt==1.17.2
google==3.0.0
googleapis-common-protos==1.70.0
plyfile==0.9
kiui==0.2.16
easydict==1.13
opentelemetry-api==1.34.1
fastapi==0.115.12
uvicorn==0.34.3
PyYAML==6.0.2
ninja==1.13.0

utils3d @ git+https://github.com/EasternJournalist/utils3d.git
trimesh

# setup for render-env
# pip install imageio==2.35.1 open3d==0.18.0 meshio==5.3.5
# pip install --no-build-isolation git+https://github.com/nerfstudio-project/gsplat.git@v1.5.3
